Flight profiles and interface improvements
Split tylo_optimizer_test into two files: single_stage_simulation and tylo_optimizer_test.

single_stage_simulation is intended to be a library of functions for simulating a craft.
-Added simulate_flight_profile to fly the craft along a pre-set flight profile (details below)
-Added optimize_engine_layout. Ingests a loadout and a flight profile, outputs a better loadout
(if it can find one).

tylo_optimizer_test is now a true example program showing instantation of a flight profile and
an engine loadout, and then feeding those into optimize_engine_layout. The example flight profile
is starting in Kerbin orbit, flying to Tylo surface, and back to Kerbin. Gravity assists are
assumed to be negligible, but you can specify correction burns too.

Flight profiles:
This is currently pretty basic. There are two types of maneuvers:
1. Vacuum landing on a planet/moon at a given altitude, and returning to orbit.
-This uses constant altitude landings / launches, and doesn't account for lofted trajectories.
Maybe this can come in a future update?
This is of the format "name altitude," for example "Mun 4000" to land at 4000m elevation on the Mun.

2. Low TWR vacuum burns (intended to be for orbital maneuvers). This leverages the use_dv function
that Arolaun added recently.
This is of the format "value", for example "200" to specify a burn of 200 m/s.

Bug fix:
Fixed a bug with selecting an engine to throttle if landing with an engine at partial throttle.
This should also slightly increase payload fraction in some situations, and greatly increase payload fraction
in some others.
